Handle "--> THIS_STATE" and -- correctly in GraphViz.
This adds a new member of EventAndSideEffects: NoTransitionEventAndSideEffects.
This name is even more long and stupid than EventAndSideEffects, and
we should change it when we figure out a better name for it.

The reason for extending EventAndSideEffects is that these are fundamentally
different kinds of event handlers and any backend will need to make
the distinction between them.