Add a semantic pass to check that all declared state names are unique.
I chose to implement this before ValidDestinationStates, as that pass was a
little too hard to wrap my head around in practice.  This was a nice middle
ground, in that it was simple, and allowed me to correct the semantic pass
interface from being just a rough draft into working code.

I also added a test for UniqueStateNames under "counterexamples".