Add a special type of "entry" state to manage state machine init.
When calling mkGraph, a transient state of type StateEntry gets placed before
all other states, and is given the node number 1; then, the user-defined start
state is connected as the transition of this StateEntry state.

In graphviz, these get drawn as filled, black, label-less circles, UML-style.
In C, these generate state variables initialized to the correct starting state.
The one catch there is that the state machine was never initialized; this
requires generating an initialize function, that calls the initial state enter
function on first run, and then peppering that throughout the events (so that
calling any event will guarantee to initialize the state machine one time, if
used single-threaded).  To do this, I modified makeEnum to allow for nameless
enums.

Also, for all functions that do not take a value, the function parameters are
now set to void for stricter type-checking.

The bulk of this work was done back in January, other than the void parameters.