Resolve #7 and #16
Two passes are added: one to generate symbol tables, and one to reduce them to
be unique and compatible across uses.  This table is then used to generate call
applications for side effects in the CStatic backend (the graphviz backend does
not yet take advantage of the symbol table).

The OverlappingInstances change to Trashcan.Graph was due to a complaint about
Monoid instances in main.