Fix #17: Exit functions are generated for terminal states.
This was solved by adding helper function finalStates, which traverses the
states, then the handlers, and decides whether that state can transition
out; if it can, it is non-final.

For prettiness' sake, I also pulled out the precedence table into a helper
function.  For functionality's sake, I inverted the logic for obtaining the
handler map, so that it can work either off of events (which CStatic wants)
or states (which finalStates wants).