Switch from the graph to the AST to construct symbol table.
This introduces the definition part of the solver, including adding Events into
the table.  Some compromises were made, namely to shove simplification into
definition at the leaves, instead of at the root, which turned out to be too
much work.  The State Monad is included even though it's not being used.