Fix #23: Optional semantic pass to strictly reject conflicting side effects
This adds the --strict flag that elaborates witohut polymorphically abstracting
bottom to void, instead passing bottom out of the typechecker where it will
fail later (in CStatic, which can't handle Tysets).

This doesn't add a semantic pass, however.  It is a bit complicated to do that
here; the solver is actually structured to report errors at instantiation and
post-canonicalization, although it doesn't yet do that either.  Baby steps.