Pull bindings out of types.
This corrects a bit of a misfire in the definition of types.  Originally, the
bindings had to be carried around with the types in order to reconcile them
later, but with the solver this hack is no longer necessary, and since it gets
in the way, good riddance!

In terms of the interface, the main change is that now the symbol table points
to both a binding and a type for a given name.

In terms of the implementation, I tried to limit the complications to a couple
sites - specifically `solve` and `defModule`.  `infer*` are about equal, but a
bit shuffled around.