Fix bug in syntax, and minor corresponding bug in parser.
This prevents, e.g., sm { * a--> b, ... }

The program abnfgen produced invalid smudge given the old abnf definition, by
concatenating states and events to their arrow/dash.  Preventing this requires
a production to break them apart; empty was sufficient for the parser, but
insufficient for the generator.  The parser still failed to accept invalid
smudge files, but for the wrong reason.