Optimization to elide exit function generation in most cases.
Although Smudge was not generating exit functions for terminal states, there
are many other cases where the exit functions did nothing but were still
generated.  This gets rid of all such cases excluding the any state, which
is harder to decide, and I don't think can cause a warning.

The impetus for this was that in states that only exited from the any state but
had no exit event, exit functions were being generated but never called.  The
two are decoupled, so in order to remove those functions, it has to go one step
further and not ever call empty exits, which can then be elided.