Close #20: Refactor C backend to simplify.
This breaks CStatic into two modules: one that translates the graph into an
intermediate representation ("Smudge IR"), and one that translates that IR into
well-structured, static-dispatch C89 just like before.

There are a few compromises that I made in the IR that will probably need to
come out with time, as well as a few remaining hacks in CStatic.  This was
mostly to preserve as much parity between the old generated output and the new,
in order to have more confidence about its correctness.

The hacks for the state name function are especially egregious; that function
was drastically different from the rest of the generated output, and required
more complex syntax that should probably remain in the individual backend, as
any given languages would likely implement it differently.

Although #20 was distant with respect to roadmap, I felt it was too difficult
to make progress with the amount of technical debt in CStatic.  Small fixes
sometimes resulted in hours of work, and it was a pain to add new features.
With this, future additions should be greatly simplified, as they will mostly
be confined to the IR.