Fix #13: Message Passing is Complicated.
This puts the wrapper into the header.  Everything now compiles and generates
substantially similar output to the manual example.  Certain deviations were
required from a practical standpoint, including exposing the structure of the
wrapper.  There is a lot of technical debt that had to be added to avoid the
need to rewrite a bunch of it again, but it's nothing insurmountable.