Progress on #65: Standardize on a properly injective mangling scheme
See the ticket for the mangling scheme.

The default case isn't beautiful, but overall this favors the identity case
while minimally punishing the simple-underscore case.

My reasoning is that mangling with a length-code-with-trailing-underscore
everywhere must be injective; you don't even need to have a join character
then.  But, if we simply included a length code before every name, it wouldn't
meet the prettiness goal.  So, removing it from one class at a time is easy:
add a join character, and then you can remove the length from all names that
don't include that character.  This gives us the first case.  Next, bump that
to a join sequence, and you can remove the length from all names that don't
include that sequence.  That gives us the second case.  Then, leading numbers
need to be disambiguated from length codes, so toss in another underscore to
the length code.

I haven't decided yet if I want to change the base for encoding the length and
special characters.  Now would be the time to do so.

This requires a proof of injectivity before I'm satisfied that it is complete.