Amendment to #65: Standardize on a properly injective mangling scheme Open #72: C Identifiers can't begin with numbers or underscores
Move the length code to the end in the general case, to reduce the number of
output identifiers with leading underscores, which is undefined in C.

A negative side-effect is that leading numbers will now generate invalid output
identifiers.  This was discussed ad-nauseum, and it was decided that it would
be better to simply fail in the backend.  Addressed by #72.