Add Traversable instances to SmudgeIR.
This one was a bit tricky, because traverse sequences.  In a strict monad like
Except that forces the whole list, which loops for infinite lists.

As a hack for now, these lists can be capped at 127.  This isn't ideal, for
three reasons:

1. It's inelegant because it hardcodes an arbitrary maximum.

2. It's inefficient because all 127 will be forced every time.

3. It's approximate because there could be an error on an unused entry.

However, it's expedient, so it wins.  In time, it might be the right choice to
move to a lazy exception model, in which case this cap can be lifted.