Progress on #72: C Identifiers can't begin with numbers or underscores
Add a flag that lets the user prefix ALL names with the namespace.

The exceptions are event and state IDs, where the prefix winds up as infix, and
event fields in the message wrapper, where the prefix is dropped.  This is due
to these respective prefixes being done very late.  Since this change is to
support injective, error-free mangling, those exceptions will not be addressed.