Close #44: GraphViz - concatenate parallel edges.
To start, concatenate the labels with an interstitial blank line.

This runs into an occasional issue with text spacing across edges, but that
could only be handled by GraphViz, so it's not clear how that could be resolved.
As this is much better as compared with the previous status quo with respect to
the difficulty of understanding, it'll have to do for now.