Progress on #38: Event logging for debugging
Listing only the events to log, generated for current design using all events.
Consume that structure in IR in a minimalistic way by logging in only one place
if all events are enabled, not at all if no events are enabled, and otherwise
according to a switch statement for the enabled events.