Add the beginnings of a debian package build.
I'm including the example files generated by dh, because at some point
we should probably make this a for-real package that could actually go
into Debian proper. Until then, though, this will build a binary .deb
file that we can include in releases to make things a bit easier for
users to install Smudge on debian-based distros.

To see what has yet to be done, run lintian on the generated .deb
file. It's very not compliant with Debian policy.