Redefine pragma syntax.
The parser was choking on trailing whitespace for arg-less pragmas.  Handling
that led to the question of how best to treat leading and trailing whitespace in
arguments.  Significant whitespace being uncommon, it was decided to trim.

The argument was also made that a likely source of confusion would be for people
to place a space before the optional equal sign - this would be very bad, as it
would cause the equal sign to be parsed as part of the argument.  This was
compelling enough that removing the equal sign seemed like the right choice.

Both of these complemented each other, since the equal sign was only useful to
enable leading and trailing whitespace inside arguments.