Fix #21: Reject event cycles.
The solution breaks a machine into basic blocks---uninterrupted runs of
self-events.  Looking at the post-conditions of the blocks, Smudge may determine
that a block certainly does or doesn't terminate.

However, Smudge is a queue machine, and as such is Turing-complete.  There is no
way to do this for all cases without attempting to solve the halting problem.