Break syntax definition apart into lexer and parser.
This has the advantage of being more modular, and easier to modify in the
future, along with cleaning up the parser by pulling out whitespace, at the
expense of verbosity.  One quirk in the split is that pragmas are hard to lex,
but their placement is restricted, so they are parsed directly.

There are two key changes in behavior that were not kept backwards-compatible.
First, `any` is no longer restricted at the parser level, but is intended to be
caught afterwards.  Second, whitespace is now allowed around the dot inside a
qualified event.