Changes to the definition to reflect the grammar rewrite.
This splits out the ASCII character classes, which were hard to translate, and
didn't add much anyhow as compared to the rest of the lexer, which was more
valuable to include.  It also restructures 'state' and 'state-def' to fit on a
single line, but coincidentally led to more natural definitions.