From-scratch rewrite of parser, now with a lexer.
The old Smudge parser was completely removed.

The new one mirrors the syntax breakup, although with a few deviations for the
sake of convenience (e.g., using the more inclusive definition of visible
characters instead of the old 'symbols' production).

To account for the change in the behavior of `any`, a new rewriting pass and two
new semantic passes were added.

Another ramification of the break up that went un-noticed is that identifiers
beginning with `--` are no longer valid when unquoted.  The lexer is forced to
choose, and as such chooses DASH instead.  The benefit or detriment of this is
up for debate.

A few tweaks were necessary to the syntax definition, namely the addition of the
sequence '\r\n' as a newline, removing empty space from individual tokens,
reordering `token` in precedence order, and a special category for visible
characters other than the double quote (without which quoted strings cannot be
terminated).  This last note raises the point that the double quote is now the
unique character (other than control codes) that cannot be included in a quoted
identifier.