Fix #42: Add full unicode support for quoted identifiers.
Expand parser to consume unicode characters not only within quoted identifiers,
but also alpha and numeric unicode in unquoted identifiers.

Emit these characters in the C unparser as hex-escaped utf-8 encoded bytes.  The
unicode escape was not used so as to remain C89 compliant (as these were added
in C99).